Add performance_rating_threshold attribute to the Staff class and refactor set_performance_rating method
