+++ apprentice_teacher.rb
@@ -1,4 +1,3 @@
-require_relative 'person'
 require_relative 'staff'
 
 class ApprenticeTeacher < Staff
@@ -6,6 +5,7 @@ class ApprenticeTeacher < Staff
   def initialize(options={})
     super
     @target_raise = 800
+    @performance_rating_threshold = 80
   end
 
   def teach_stuff
@@ -13,17 +13,6 @@ class ApprenticeTeacher < Staff
     response += "Listen, class, this is how everything works. "
     response += "*drops crazy knowledge bomb* "
     response += "... You're welcome."
-    response
-  end
-
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 80
-      response = "Yay, I'm a great employee!"
-      receive_raise(@target_raise)
-    else
-      response = "Oh, well -- thanks to this actionable, specific, and kind feedback, I'll do better next time."
-    end
     response
   end
