+++ staff.rb
@@ -2,7 +2,7 @@ require_relative 'person'
 
 class Staff < Person
 
-  attr_reader :salary, :target_raise
+  attr_reader :salary, :target_raise, :performance_rating_threshold
 
   def initialize(options={})
     super
@@ -21,6 +21,17 @@ class Staff < Person
 
   def receive_raise(raise)
     @salary += raise
+  end
+
+  def set_performance_rating(rating)
+    response = ""
+    if rating > @performance_rating_threshold
+      receive_raise(@target_raise)
+      response = "Yay, I'm a great employee!"
+    else
+      response = "Oh, well -- thanks to this actionable, specific, and kind feedback, I'll do better next time."
+    end
+    response
   end
 
 end
