+++ teacher.rb
@@ -1,4 +1,3 @@
-require_relative 'person'
 require_relative 'staff'
 
 class Teacher < Staff
@@ -7,6 +6,7 @@ class Teacher < Staff
   def initialize(options={})
     super
     @target_raise = 1000
+    @performance_rating_threshold = 90
   end
 
   def teach_stuff
@@ -17,15 +17,4 @@ class Teacher < Staff
     response
   end
 
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 90
-      receive_raise(@target_raise)
-      response = "Yay, I'm a great employee!"
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
-    end
-    response
-  end
 end
